Updated all R scripts to include automatic downloading and unzipping of data if the required file is not found in the working directory.

+++ plot1.R
@@ -1,10 +1,16 @@
+# change working directory
 setwd("c:/users/liam/documents/r/datascience/exploratory data analysis")
 
-# download zipped data file, extract the required data file and remove zipped file
-temp <- tempfile()
-download.file("https://d396qusza40orc.cloudfront.net/exdata%2Fdata%2Fhousehold_power_consumption.zip", temp)
-electric.power.consumption = read.csv(unz(temp, "household_power_consumption.txt"), header=TRUE, sep=";", na.strings="?", colClasses=c("character","character","numeric","numeric","numeric","numeric","numeric","numeric","numeric"))
-unlink(temp)
+# if necessary, download zipped data file, extract the required data file and remove zipped file
+if (!file.exists("household_power_consumption.txt")) {
+    temp <- tempfile()
+    download.file("https://d396qusza40orc.cloudfront.net/exdata%2Fdata%2Fhousehold_power_consumption.zip", temp)
+    unz(temp, "household_power_consumption.txt")
+    unlink(temp)    
+}
+
+# load the semi-colon delimited data into a data table
+electric.power.consumption = read.csv(, header=TRUE, sep=";", na.strings="?", colClasses=c("character","character","numeric","numeric","numeric","numeric","numeric","numeric","numeric"))
 
 # extract the two days of interest
 electric.power.consumption = electric.power.consumption[as.character(electric.power.consumption$Date) == "1/2/2007"
